# server/docs/swagger.yaml
definitions:
  calendar.CalendarEventsWithError:
    properties:
      calendarEvents:
        items:
          $ref: '#/definitions/models.CalendarEvent'
        type: array
      error:
        type: error
    type: object
...
  models.AppleCalendarAuth:
    type: object
...
  models.BufferTimeOptions:
    properties:
      enabled:
        type: boolean
      time:
        type: integer
    type: object
  models.CalendarAccount:
    properties:
      appleCalendarAuth:
        $ref: '#/definitions/models.AppleCalendarAuth'
        type: object
      calendarType:
        type: string
      email:
        description: Email is required for all calendar accounts
        type: string
      enabled:
        type: boolean
      oAuth2CalendarAuth:
        $ref: '#/definitions/models.OAuth2CalendarAuth'
        type: object
      picture:
        type: string
      subCalendars:
        additionalProperties:
          $ref: '#/definitions/models.SubCalendar'
        type: object
    type: object
  models.CalendarEvent:
    properties:
      allDay:
        description: Whether the event is an all day event
        type: boolean
      calendarId:
        type: string
      endDate:
        type: string
      free:
        description: Whether the user is free during this event
        type: boolean
      id:
        type: string
      startDate:
        type: string
      summary:
        type: string
    type: object
  models.CalendarOptions:
    properties:
      bufferTime:
        $ref: '#/definitions/models.BufferTimeOptions'
        type: object
      workingHours:
        $ref: '#/definitions/models.WorkingHoursOptions'
        type: object
    type: object
  models.CalendarType:
    type: string
...
  models.EventType:
    type: string
...
  models.OAuth2CalendarAuth:
    type: object
...
  models.SignUpBlock:
    properties:
      _id:
        type: string
      capacity:
        type: integer
      endDate:
        type: string
      name:
        type: string
      startDate:
        type: string
    type: object
...
  models.SubCalendar:
    properties:
      enabled:
        type: boolean
      name:
        type: string
    type: object
  models.User:
    properties:
      _id:
        description: Profile info
        type: string
      calendarAccounts:
        additionalProperties:
          $ref: '#/definitions/models.CalendarAccount'
        description: |-
          CalendarAccounts is a mapping from {`email_CALENDARTYPE` => CalendarAccount} that contains all the
          additional accounts the user wants to see google calendar events for
        type: object
      calendarOptions:
        $ref: '#/definitions/models.CalendarOptions'
        description: Calendar options
        type: object
      email:
        type: string
      firstName:
        type: string
      hasCustomName:
        description: Whether the user has set a custom name for themselves, i.e. don't
          change their name when they sign in
        type: boolean
      isPremium:
        type: boolean
      lastName:
        type: string
      numEventsCreated:
        type: integer
      picture:
        type: string
      primaryAccountKey:
        description: The calendarAccountKey of the account the user first signed in
          with
        type: string
      stripeCustomerId:
        description: Stripe customer ID
        type: string
      timezoneOffset:
        type: integer
    type: object
  models.WorkingHoursOptions:
    properties:
      enabled:
        type: boolean
      endTime:
        type: number
      startTime:
        type: number
    type: object
...
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for Schej.it!
  license: {}
  title: Schej.it API
  version: "1.0"
paths:
...
  /analytics/user/{email}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Gets the user by email
      tags:
      - analytics
...
  /events:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing info about the event to create
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              attendees:
                items:
                  type: string
                type: array
              blindAvailabilityEnabled:
                type: boolean
              dates:
                items:
                  type: string
                type: array
              daysOnly:
                type: boolean
              duration:
                type: number
              isSignUpForm:
                type: boolean
              name:
                type: string
              notificationsEnabled:
                type: boolean
              remindees:
                items:
                  type: string
                type: array
              sendEmailAfterXResponses:
                type: integer
              signUpBlocks:
                items:
                  $ref: '#/definitions/models.SignUpBlock'
                type: array
              timeIncrement:
                type: integer
              type:
                $ref: '#/definitions/models.EventType'
              when2meetHref:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - type: object
            - properties:
                eventId:
                  type: string
              type: object
      summary: Creates a new event
      tags:
      - events
...
  /user/add-apple-calendar-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing the email and app password of the apple account
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              email:
                type: string
              password:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Adds an apple calendar account
      tags:
      - user
  /user/add-google-calendar-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing the Google authorization code and scope
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              code:
                type: string
              scope:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Adds a new calendar account
      tags:
      - user
  /user/add-outlook-calendar-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing the Outlook authorization code and scope
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              code:
                type: string
              scope:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Adds a new outlook calendar account
      tags:
      - user
  /user/calendar-options:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Object containing the updated options
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              bufferTime:
                $ref: '#/definitions/models.BufferTimeOptions'
              workingHours:
                $ref: '#/definitions/models.WorkingHoursOptions'
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Updates the user's calendar options
      tags:
      - user
  /user/calendars:
    get:
      description: Gets the user's calendar events between "timeMin" and "timeMax"
      parameters:
      - description: Lower bound for event's start time to filter by
        in: query
        name: timeMin
        required: true
        type: string
      - description: Upper bound for event's end time to filter by
        in: query
        name: timeMax
        required: true
        type: string
      - description: Comma separated list of accounts to fetch calendar events from
        in: query
        name: accounts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/calendar.CalendarEventsWithError'
            type: object
      summary: Gets the user's calendar events
      tags:
      - user
...
  /user/events/{eventId}/set-folder:
    post:
      consumes:
      - application/json
      parameters:
      - description: The ID of the event to set the folder for
        in: path
        name: eventId
        required: true
        type: string
      - description: The ID of the folder to set the event to
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              folderId:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Sets the folder for the specified event
      tags:
      - user
...
  /user/name:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Object containing the updated name
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              firstName:
                type: string
              lastName:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Updates the user's name
      tags:
      - user
...
  /user/remove-calendar-account:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Object containing the email + type of the calendar account to
          remove
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              calendarType:
                $ref: '#/definitions/models.CalendarType'
              email:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Removes an existing calendar account
      tags:
      - user
  /user/searchContacts:
    get:
      parameters:
      - description: Query to search for
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Searches the user's contacts based on the given query
      tags:
      - user
  /user/toggle-calendar:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email of calendar account and whether to enable it
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              calendarAccountKey:
                type: string
              enabled:
                type: boolean
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Toggles whether the specified calendar is enabled or disabled for the
        user
      tags:
      - user
  /user/toggle-sub-calendar:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email of calendar account, the sub calendar id, and whether to
          enable it
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              calendarAccountKey:
                type: string
              enabled:
                type: boolean
              subCalendarId:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Toggles whether the specified sub-calendar is enabled or disabled for
        the user
      tags:
      - user
  /users:
    get:
      parameters:
      - description: Search query matching users' names/emails
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: An array of user profile objects
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Returns users that match the search query
      tags:
      - users
  /users/{userId}:
    get:
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A user profile object
          schema:
            $ref: '#/definitions/models.User'
      summary: Returns the user by their user id
      tags:
      - users
swagger: "2.0"